Pass message updates first to MessageDb and then to MessageModel
This makes it possible to update messages in MessageModel directly after
adding a message via MessageDb.
It is needed because of the asynchronous processing of messages.

An example are delivery state updates which are fixed now.
The delivery state could not be updated because the message had not yet
been added to the loaded messages at the time of updating it.