Implement read markers
Read markers are implemented as specified by XEP-0333: Chat Markers.
Only the "displayed" part of the XEP is implemented.

Read markers are sent when a message marked as "markable" is received
and displayed to the user or marked as read via a notification's action.
When read markers are received, a text is shown in the corresponding
chat.

Additionally, the counter for unread messages is decreased and
notifications are closed if read markers for messages from a contact are
received from another own client.
That snychronizes the state of read messages between multiple endpoints.

Closes #299, #367

Co-authored-by: Bhavy Airi <airiragahv@gmail.com>