Add support for blocking JIDs
This implements XEP-0191: Blocking Command

Blocked JIDs are stored in the database to be viewed when not connected.
Options for blocking/unblocking are not provided when not connected.

Roster items can be blocked via their details.
All blocked JIDs can be managed and new ones blocked via the account details.

Users (bare JIDs), servers (domains), devices of users (full JIDs) and
devices on servers (domains with resource parts) can be blocked.