Fetch messages via MAM after the last message instead of starting with it
That is done by passing the last message's stanza ID to be used as the
Result Set Management's 'after' element.

Formerly, the last message's timestamp was passed via MAM's 'start' field.
That resulted in the same message being fetched.
Even multiple already locally stored messages could be fetched because of
time differences between the server and the client.