MessageHandler: Fix fetching initial MAM message for each roster item
When the user logs in with an account via Kaidan for the first time,
an inital message for each roster item is fetched from the server via MAM.

The following process is now done to ensure that a message with a body or
shared file fetched and stored as the initial message.
The latest message for the corresponding chat is fetched from the server.
If that is a message without a body, the next older message is fetched.
That is repeated until the first message with a body is fetched.
The message is stored as the initial message and its body/file is displayed
as the latest message for the corresponding chat within the RosterPage.

Closes #463