Add JID to roster on receiving presence subscription request
Formerly, a contact could be lost if no message was exchanged and the
inline notification was closed without accepting the presence subscription
request.

The new approach adds new contacts to the roster on receiving a presence
subscription request.
In case the contact sent a text within the request, that text is added as
the intial message resulting in a message notification.
Otherwise, a default text is shown in the notification.
The presence subscription request can be accepted later via a chat hint.