Fix resetting current chat
If a contact was removed and added again without switching to
another chat in between, the corresponding chat page could not be
opened again because that chat was still set as the current one.
The current chat is now always reset when the chat page is closed
(e.g., because of a contact removal).