Always allow to authenticate encryption keys
The buttons to open the key authentication pages are now visible even
if all keys are authenticated or none available/known.
That enables other devices to authenticate the user device's key even if
the user's device already authenticated the other devices.
That makes it also possible to authenticate keys that are unknown to the
user's client (e.g., keys that the client has not received yet because it
was offline while contacts published new keys).

Additionally, descriptions are added in order to explain what the current
state is and what can be done to improve it.

When the user tries to authenitcate an unknown key, it is not rejected
anymore.