main: Fix setting styles
The Qt Quick Controls style is set before instantiating the app.
That fixes the Qt Quick Controls style not being recognized by
Kirigami.

The fallback icon theme is set instead of the main icon theme if
no fallback is set.
That fixes displaying the system's icons if available.
Otherwise, the fallback icon theme is used.