FutureUtils: Fix possible wrong thread watcher initialization
Depending on how things are called, the following QObject error can
occur: "Cannot create children for a parent that is in a different
thread".
This is due to the watcher being created in the calling thread which
can have a different context.
Thus, it should be moved to the target thread.