Add custom log message handler to filter out specific messages
Log messages containing the strings defined in the environment
variable KAIDAN_LOG_FILTER and separated by semicolons are filtered
out of the log.

Ignored QML enum warnings are filtered out of the log.
Enforcing enum classes results in QML warnings because the enums
are exported both globally and within the enum name in that case.