Apply consistent criteria for all message corrections
This applies the criteria for outgoing message corrections to
incoming message corrections.
That results in consistent behavior when to allow message
corrections, i.e., when to store/display them as such.
Incoming message corrections that do not meet the criteria are
added as new messages.

Criteria:
* A message is suitable for correction only if it is among the 10
most recent messages.
* A message is suitable for correction only if it has not been sent
earlier than 2 days ago.

All criteria must be met for a valid message correction.