Show input field invalid hint manually and position cursor on focus
Formerly, the hint for invalid input of an input field was shown
once the field lost its focus.
That resulted in various problems.
E.g., the invalid hints of the JID and password fields were shortly
shown after clicking a button on the start page to begin the
account registration.

The invalid hints are now only shown after pressing the Enter
button or the corresponding confirmation button (e.g., the login
button).

In case of an invalid input, the input is not selected to replace
it completely.
Instead, the cursor is moved to the end of the input in order to
modify specific parts of it.