RosterDb: Fix retrieving name of last message's sender in group chat
Formerly, the name of the last message's sender in a group chat was
determined by using the name of the first roster item with the same
JID found in the database.
But that item can have a different name if another account has a
different name for the same contact.
Thus, the account JID is now checked as well.