Add provider for all images
The provider produces images for many cases.
The images are requested by URLs.
That makes it easy to embed various image sources in QML Image
items.

The functionality of the former image provider for Bits of Binary
data is integrated in the new central image provider.

The following cases are covered:

* Local files: Use KDE Thumbnailer (which stores thumbnails in
a cache directory)
* Icons: Use QIcon::fromTheme()
* QR codes: Use KDE Prison
* Base64-encoded images: Use QImage::fromData()