Add support for offline media sharing queue
This allows to add messages that include files while not being
connected to the server.
Once connected and an upload service found, the files are uploaded.
Once all files are uploaded, the message is sent.

This also restarts uploads again once connected if they were
canceled on disconnecting from the server.

This needs QXmpp 1.13.